requirements.txt
faiss-cpu
sentence-transformers
openai
streamlit
chromadb
numpy
